#104 - billing address - move form and whole functionality to separate component
