+++ view/frontend/web/js/components/steps/ThePaymentStep.vue
@@ -16,128 +16,22 @@
       text="My billing and shipping address are the same"
     />
     <AddressData
-      v-if="billingAndShippingAddressTheSame"
+      v-if="billingAndShippingAddressTheSame && billingAddress !== null"
       :address="billingAddress"
     />
     <div v-else>
-      <form
-        v-if="editBillingAddress"
-        class="billing-address__form"
-      >
-        <div>
-          <BaseInput
-            v-model="$v.address.firstname.$model"
-            :validation="$v.address.firstname"
-            label="First name"
-            name="firstname"
-          />
-          <BaseInput
-            v-model="$v.address.lastname.$model"
-            :validation="$v.address.lastname"
-            label="Last name"
-            name="lastname"
-          />
-          <BaseInput
-            v-model="$v.address.telephone.$model"
-            :validation="$v.address.telephone"
-            label="Phone Number"
-            name="telephone"
-            type="tel"
-          />
-          <BaseInput
-            v-model="$v.address.street0.$model"
-            :validation="$v.address.street0"
-            label="Street Address"
-            name="street0"
-          />
-          <BaseInput
-            v-model="address.street1"
-            label="Street Address"
-            name="street1"
-          />
-          <div>
-            <label for="country">
-              Select Country
-            </label>
-            <multiselect
-              id="country"
-              v-model="$v.address.country_id.$model"
-              :options="countries"
-              :allow-empty="false"
-              :show-labels="false"
-              name="country"
-              label="label"
-              placeholder="Select country"
-            />
-            <span
-              v-if="
-                $v.address.country_id.$error
-                  && !$v.address.country_id.required
-              "
-            >
-              This field is required!
-            </span>
-          </div>
-          <BaseInput
-            v-model="$v.address.city.$model"
-            :validation="$v.address.city"
-            label="City"
-            name="city"
-          />
-          <BaseInput
-            v-model="$v.address.postcode.$model"
-            :validation="$v.address.postcode"
-            label="Zip/Postal Code"
-            name="postcode"
-          />
-          <BaseInput
-            v-if="!regions.length"
-            v-model="$v.address.region.$model"
-            :validation="$v.address.region"
-            label="State/Province"
-            name="region"
-          />
-          <div v-if="regions.length">
-            <label for="region_id">
-              Select State/Province
-            </label>
-            <multiselect
-              id="region_id"
-              v-model="$v.address.region_id.$model"
-              :validation="$v.address.region_id"
-              :options="regions"
-              :allow-empty="false"
-              :show-labels="false"
-              name="region_id"
-              label="label"
-              placeholder="Select State/Province"
-            />
-            <span
-              v-if="
-                $v.address.region_id.$error
-                  && !$v.address.region_id.required
-              "
-            >
-              This field is required!
-            </span>
-          </div>
-          <BaseInput
-            v-model="address.company"
-            label="Company"
-            name="company"
-          />
-          <BaseButton
-            text="Save Address"
-            @click.native="saveAddress"
-          />
-          <BaseButton
-            text="Cancel"
-            @click.native="toggleBillingAddress"
-          />
-        </div>
-      </form>
+      <template v-if="editBillingAddress">
+        <BillingAddressForm @hideAddressForm="editBillingAddress = false"/>
+        <BaseButton
+          text="Cancel"
+          @click.native="toggleBillingAddress()"
+        />
+      </template>
       <div v-else>
-        <AddressData :address="newBillingAddress" />
+        <AddressData
+          v-if="newBillingAddress !== null"
+          :address="newBillingAddress"
+        />
         <BaseButton
           text="Edit Address"
           @click.native="editAddress"
@@ -166,11 +60,8 @@
 import AddressData from '../AddressData.vue'
 import BaseButton from '../BaseButton.vue'
 import BaseCheckbox from '../BaseCheckbox.vue'
-import BaseInput from '../BaseInput.vue'
+import BillingAddressForm from '../BillingAddressForm.vue'
 import DiscountCodeForm from '../DiscountCodeForm.vue'
-import { required, requiredIf } from 'vuelidate/lib/validators'
-import countries from '../../data/countries.json'
-import Multiselect from 'vue-multiselect'
 import PaymentMethods from '../PaymentMethods.vue'
 
 export default {
@@ -178,27 +69,12 @@ export default {
     AddressData,
     BaseButton,
     BaseCheckbox,
-    BaseInput,
+    BillingAddressForm,
     DiscountCodeForm,
-    Multiselect,
     PaymentMethods
   },
   data () {
     return {
-      address: {
-        firstname: '',
-        lastname: '',
-        telephone: '',
-        street0: '',
-        street1: '',
-        country_id: '',
-        city: '',
-        postcode: '',
-        region: '',
-        region_id: '',
-        company: ''
-      },
-      countries,
       billingAndShippingAddressTheSame: true,
       editBillingAddress: true,
       selectedPaymentMethod: null,
@@ -206,45 +82,7 @@ export default {
       loader: false
     }
   },
-  validations: {
-    address: {
-      firstname: {
-        required
-      },
-      lastname: {
-        required
-      },
-      telephone: {
-        required
-      },
-      street0: {
-        required
-      },
-      country_id: {
-        required
-      },
-      city: {
-        required
-      },
-      postcode: {
-        required
-      },
-      region: {
-        required: requiredIf(function () {
-          return this.regions.length === 0
-        })
-      },
-      region_id: {
-        required: requiredIf(function () {
-          return this.regions.length > 0
-        })
-      }
-    }
-  },
   computed: {
-    cartId () {
-      return this.$store.getters.cartId
-    },
     billingAddress () {
       return this.$store.state.billingAddress
     },
@@ -272,16 +110,6 @@ export default {
   methods: {
     toggleBillingAddress () {
       this.billingAndShippingAddressTheSame = !this.billingAndShippingAddressTheSame
-    },
-    saveAddress () {
-      this.$v.$touch()
-      if (!this.$v.$invalid) {
-        this.editBillingAddress = false
-        this.$store.commit(
-          'setAddress',
-          { type: 'newBillingAddress', address: this.address }
-        )
-      }
     },
     editAddress () {
       this.editBillingAddress = true
